+++ cocode/src/main/java/dgsw/b1cami/cocode/Repository/LunchRepository.java
@@ -12,13 +12,13 @@ import java.util.ArrayList;
 import java.util.Optional;
 
 @Repository
-public interface LunchRepository extends JpaRepository<Lunch, Integer> {
+public interface LunchRepository extends JpaRepository<Lunch, Long> {
 
-    Optional<Lunch> findByLunchId(Integer lunchId);
+    Optional<Lunch> findByLunchId(Long lunchId);
 
     @Transactional
     @Modifying
     @Query(value = "SELECT * FROM lunch WHERE lunch_id >= (:lunchCount - 20) AND lunch_id < :lunchCount ORDER BY lunch_id DESC", nativeQuery = true)
-    ArrayList<Lunch> findLunches(@Param("lunchCount") Integer lunchCount);
+    ArrayList<Lunch> findLunches(@Param("lunchCount") Long lunchCount);
 
 }
